requirements.txt
Adafruit-DHT==1.4.0
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.5.0
MarkupSafe==1.1.1
paho-mqtt==1.4.0
pytz==2019.1
redis==3.2.1
uWSGI==2.0.18
vine==1.3.0
Werkzeug==0.15.2
